refactor react base project to react-redux
I had a bug in the original react (without redux) project, when the user press on many floors at the same time the state is overridden.
Javascript doesn't have "lock" statement (it is one threaded) so i decided to try refactor the code to work with redux.
I managed to implement all the "Calling an elevator" and "Elevator reached the floor" parts except the audio part, but i still have bugs - some of the dispatches are being called multiple times when they only need to be called once (maybe the setTimeout is the problem, i need to keep debugging)