finished requirements, bug fixed
I decided to refactor the code a little bit-
every call for elevator will now enter the queue.
whenever the queue is updated, the useEffect with the queue dependency, will dispatch an "elevator call" for the first value in the queue.
This way, all calls are being saved and there are no overrides.
Also, the last bug was some dispatches was being dispatched twice.
That was because the condition "if(elevatorInfo.isMoving)" in the elevator component  was not "strong" enough-  isMoving changes to false only after the 2 seconds delay.
I also re-arranged the files and folders to be more organized.
I still have left the stopwatch feature, but i will implement it with the bonus part