final bonus updates !
1. i decided to give one more try to the stop watch.
 I updated to stopWatch state to be an array of objects like so:
{
      elevatorNumber: -1,
      start: false,
      pickUpTime: 0,
 }
I also added a timer to the floor object.
Once an elevator call is being dispatched, the timer is updated to preformence.now(), and the "start" value of  stopWatch (in this floor) is set to true.
Then a stop watch will appear on the right on the calling floor.
When there is an available elevator, the time from the call to this point will be calculated and "pickUpTime" will be updated accordingly.
Then, the stop watch will appear on the available elevator column (on the calling floor row). This is possible because of a condition depending on "elevatorNumber" value.

2. i added a button to go back to settings, from the main page.

3. there was an error about react stack overload, because the "handleElevatorCall" was dispatched many times (in the Building component). So, i added another condition before dispatching